igh so a lot of stuff
So, made a common user file where all the routes for users are stored n also made a route to delete a users data
