+++ go.mod
@@ -13,9 +13,9 @@ require (
 )
 
 require (
-	github.com/joho/godotenv v1.5.1 // indirect
+	github.com/joho/godotenv v1.5.1 // direct
 	github.com/philhofer/fwd v1.1.2 // indirect
-	github.com/resend/resend-go/v2 v2.10.0 // indirect
+	github.com/resend/resend-go/v2 v2.10.0 // direct
 	github.com/tinylib/msgp v1.1.8 // indirect
 )
